// app/enums/stock-categories-list.enum.ts
export enum StockCategoriesList {
  TECHNOLOGY = 'TECHNOLOGY',
  FINANCE = 'FINANCE',
  HEALTHCARE = 'HEALTHCARE',
  ENERGY = 'ENERGY',
  CONSUMER_DISCRETIONARY = 'CONSUMER_DISCRETIONARY',
  CONSUMER_STAPLES = 'CONSUMER_STAPLES',
  UTILITIES = 'UTILITIES',
  MATERIALS = 'MATERIALS',
  INDUSTRIALS = 'INDUSTRIALS',
  REAL_ESTATE = 'REAL_ESTATE',
  TELECOMMUNICATIONS = 'TELECOMMUNICATIONS',
  AGRICULTURE = 'AGRICULTURE',
  AEROSPACE = 'AEROSPACE',
}